Almost done
I had an issue whereby the score would not transfer over to the gameEndScreen only to realize that id labeling is quite crucial to properly identifying what variables get affected. Alongside this issue, I also had an issue whereby the score wouldn’t increment, I used scoreElement to fix this.

I used conditionals to increment the speed of the obstacle and used animationId && statements to determine what and how fast the obstacles would be affected.

I also wanted to create a ‘next level’ domain whereby when the player reaches the top of the screen, the score increments and the player is removed and the game is reset.

I updated the obstacle move function as a way to control the speed of the obstacle, by creating a this.speed to manipulate in the game.js file whereby I set a value for the obstacle's speed.

Probably my hardest step was to create the obstacles coming along the y axis randomly. In the obstacle js file, I created a this.pos and this.img that randomized between 0-1 and depending on what it would fall on, the result would be where the obstacle would come from. Additionally, I created another conditional to randomize the content of the obstacle, between the asteroid and alien. 

I spent the last part of the project working on styling, the game end screen, and game intro screen. I also added audio to intensify the experience as I hope the music will heighten the pressure!